game-strings: try to load base files
