+++ tools/shared/paths.py
@@ -13,7 +13,13 @@ SHARED_INCLUDE_DIR = SRC_DIR / "libtrx/include/libtrx"
 SHARED_SRC_DIR = SRC_DIR / "libtrx"
 
 
-class ProjectPaths:
+class BasePaths:
+    data_dir = Path
+    shipped_data_dir = Path
+    src_dir = Path
+
+
+class ProjectPaths(BasePaths):
     def __init__(self, folder_name: str) -> None:
         self.folder_name = folder_name
 
@@ -29,5 +35,11 @@ TR1Paths = ProjectPaths(folder_name="tr1")
 
 TR2Paths = ProjectPaths(folder_name="tr2")
 TR2Paths.progress_file = TR2Paths.docs_dir / "symbols.txt"
+
+CommonPaths = BasePaths()
+CommonPaths.data_dir = DATA_DIR / "common"
+CommonPaths.shipped_data_dir = CommonPaths.data_dir / "ship"
+CommonPaths.src_dir = SRC_DIR / "libtrx"
+CommonPaths.include_dir = SHARED_INCLUDE_DIR
 
 PROJECT_PATHS = {1: TR1Paths, 2: TR2Paths}
