+++ tools/update_game_strings
@@ -1,0 +1,175 @@
+#!/usr/bin/env python3
+import json
+import re
+from pathlib import Path
+
+from shared.json_utils import (
+    JSONPointers,
+    load_json5,
+    load_json5_from_string,
+    write_json,
+    write_json_to_string,
+)
+from shared.paths import PROJECT_PATHS, DATA_DIR, CommonPaths
+from shared.translation_utils import (
+    REVIEW_MARKER,
+    is_translation_file,
+    should_skip_object_name,
+)
+
+
+def get_objects_map(paths: list[Path]) -> dict[str, list[str]]:
+    """
+    Parse object-name definitions from names.def-like files.
+    Supports:
+      X_OBJ_NAME_DEFINE(..., "key", X_OBJ_NAMES("Name1", "Name2", ...))
+    """
+    result: dict[str, list[str]] = {}
+    for path in paths:
+        for line in path.read_text().splitlines():
+            text = line.strip()
+            match = re.match(
+                r'^X_OBJ_NAME_DEFINE\(\s*\w+\s*,\s*"([^"]+)"\s*,\s*X_OBJ_NAMES\(\s*(.+?)\s*\)\s*\)$',
+                text,
+            )
+            if match:
+                key = match.group(1)
+                names_list = re.findall(r'"([^"]+)"', match.group(2))
+                result.setdefault(key, []).extend(names_list)
+    return result
+
+
+def get_strings_map(paths: list[Path]) -> dict[str, str]:
+    result: dict[str, str] = {}
+    for path in paths:
+        for line in path.read_text().splitlines():
+            if match := re.match(
+                r'^\w+DEFINE\((\w+),\s*"([^"]+)"\)$', line.strip()
+            ):
+                result[match.group(1)] = (
+                    match.group(2).replace("\\n", "\n").replace("\\\\", "\\")
+                )
+    return result
+
+
+def postprocess_object_strings(
+    content: str,
+    object_names_map: dict[str, list[str]],
+) -> str:
+    content_dict = load_json5_from_string(content)
+    content_dict["objects"] = {
+        key: {"name": names[0]} if len(names) == 1 else {"names": names}
+        for key, names in object_names_map.items()
+    }
+    return write_json_to_string(content_dict) + "\n"
+
+
+def postprocess_game_strings(
+    content: str,
+    game_strings_map: dict[str, str],
+) -> str:
+    content_dict = load_json5_from_string(content)
+    content_dict["game_strings"] = dict(
+        sorted(game_strings_map.items(), key=lambda kv: kv[0])
+    )
+    return write_json_to_string(content_dict) + "\n"
+
+
+class ReviewTracker:
+    def __init__(self) -> None:
+        self.changed_ptrs: dict[Path, list[str]] = {}
+
+    def track_change(
+        self, eng_path: Path, old_content: str, new_content: str
+    ) -> None:
+        """Determine which entries changed for later marking in translations."""
+        old_data = load_json5_from_string(old_content)
+        new_data = load_json5_from_string(new_content)
+        old_ptr = JSONPointers(old_data)
+        new_ptr = JSONPointers(new_data)
+        self.changed_ptrs[eng_path] = [
+            ptr
+            for ptr in new_ptr
+            if new_ptr.get(ptr) != old_ptr.get(ptr)
+            and not should_skip_object_name(ptr, new_data)
+        ]
+
+    def mark_for_review(self) -> None:
+        """Mark changed entries in translation files with REVIEW_MARKER."""
+        for eng_path, ptrs in self.changed_ptrs.items():
+            for trans_path in eng_path.parent.glob(f"{eng_path.stem}-*.json5"):
+                if not is_translation_file(trans_path):
+                    continue
+                trans_data = load_json5(trans_path)
+                trans_ptr = JSONPointers(trans_data)
+                for ptr in ptrs:
+                    trans_ptr[ptr] = REVIEW_MARKER
+                write_json(trans_path, trans_data)
+
+
+def process_game_strings(
+    game_strings_def_path: Path, target_dir: Path
+) -> None:
+    game_strings_map = get_strings_map([game_strings_def_path])
+
+    # Update English base files
+    review_tracker = ReviewTracker()
+    for path in target_dir.rglob("*strings.json*"):
+        # Skip translation files; handle only base (English) files here
+        if is_translation_file(path):
+            continue
+
+        old_content = path.read_text()
+        new_content = old_content
+        new_content = postprocess_game_strings(new_content, game_strings_map)
+        if new_content != old_content:
+            # Write updated English file
+            path.write_text(new_content)
+            review_tracker.track_change(path, old_content, new_content)
+
+    review_tracker.mark_for_review()
+
+
+def process_object_names(
+    object_names_def_path: Path, target_dir: Path
+) -> None:
+    object_names_map = get_objects_map([object_names_def_path])
+    assert object_names_map
+
+    review_tracker = ReviewTracker()
+
+    # Update English base files
+    changed_ptrs: dict[Path, list[str]] = {}
+    for path in target_dir.rglob("*strings.json*"):
+        # Skip translation files; handle only base (English) files here
+        if is_translation_file(path):
+            continue
+        old_content = path.read_text()
+        new_content = old_content
+        new_content = postprocess_object_strings(new_content, object_names_map)
+        if new_content != old_content:
+            # Write updated English file
+            path.write_text(new_content)
+            review_tracker.track_change(path, old_content, new_content)
+
+    review_tracker.mark_for_review()
+
+
+def main() -> None:
+    process_game_strings(
+        CommonPaths.include_dir / "game/game_string.def",
+        CommonPaths.shipped_data_dir,
+    )
+    for game, project_paths in PROJECT_PATHS.items():
+        process_game_strings(
+            project_paths.src_dir / "game/game_string.def",
+            project_paths.shipped_data_dir,
+        )
+        process_object_names(
+            CommonPaths.include_dir / f"game/objects/names_tr{game}.def",
+            project_paths.shipped_data_dir,
+        )
+
+
+if __name__ == "__main__":
+    main()
