--- tools/update_gameflow
@@ -1,150 +1,0 @@
-#!/usr/bin/env python3
-import json
-import re
-from pathlib import Path
-
-from shared.paths import PROJECT_PATHS, SHARED_INCLUDE_DIR
-from shared.translation_utils import (
-    REVIEW_MARKER,
-    JSONPointers,
-    is_translation_file,
-    load_json5,
-    load_json5_from_string,
-    should_skip_object_name,
-    write_json,
-)
-
-
-def get_objects_map(paths: list[Path]) -> dict[str, list[str]]:
-    """
-    Parse object-name definitions from names.def-like files.
-    Supports:
-      X_OBJ_NAME_DEFINE(..., "key", X_OBJ_NAMES("Name1", "Name2", ...))
-    """
-    result: dict[str, list[str]] = {}
-    for path in paths:
-        for line in path.read_text().splitlines():
-            text = line.strip()
-            match = re.match(
-                r'^X_OBJ_NAME_DEFINE\(\s*\w+\s*,\s*"([^"]+)"\s*,\s*X_OBJ_NAMES\(\s*(.+?)\s*\)\s*\)$',
-                text,
-            )
-            if match:
-                key = match.group(1)
-                names_list = re.findall(r'"([^"]+)"', match.group(2))
-                result.setdefault(key, []).extend(names_list)
-    return result
-
-
-def get_strings_map(paths: list[Path]) -> dict[str, str]:
-    result: dict[str, str] = {}
-    for path in paths:
-        for line in path.read_text().splitlines():
-            if match := re.match(
-                r'^\w+DEFINE\((\w+),\s*"([^"]+)"\)$', line.strip()
-            ):
-                result[match.group(1)] = match.group(2)
-    return result
-
-
-def postprocess_game_strings(
-    content: str,
-    object_names_map: dict[str, list[str]],
-    game_strings_map: dict[str, str],
-) -> str:
-    max_key_length = (
-        max(len(json.dumps(key)) for key in object_names_map.keys()) + 2
-    )
-    indent = " " * 8
-    content = re.sub(
-        r'^(    "objects": {)[\S\s]*?^(    })',
-        '    "objects": {\n'
-        + "\n".join(
-            (
-                f"{indent}{f'{json.dumps(key)}: '.ljust(max_key_length)}"
-                f"{json.dumps({'name': names[0]})},"
-                if len(names) == 1
-                else f"{indent}{f'{json.dumps(key)}: '.ljust(max_key_length)}"
-                f"{json.dumps({'names': names})},"
-            )
-            for key, names in object_names_map.items()
-        )
-        + "\n    }",
-        content,
-        flags=re.M | re.DOTALL,
-    )
-
-    content = re.sub(
-        r'^(    "game_strings": {)[\S\s]*?^(    })',
-        '    "game_strings": {\n'
-        + "\n".join(
-            f"        {json.dumps(key)}: {json.dumps(value)},"
-            for key, value in sorted(
-                game_strings_map.items(), key=lambda kv: kv[0]
-            )
-        )
-        + "\n    }",
-        content,
-        flags=re.M | re.DOTALL,
-    )
-    return content
-
-
-def process(game: int) -> None:
-    game_string_def_paths = [
-        SHARED_INCLUDE_DIR / "game/game_string.def",
-        PROJECT_PATHS[game].src_dir / "game/game_string.def",
-    ]
-    object_names_def_path = (
-        SHARED_INCLUDE_DIR / f"game/objects/names_tr{game}.def"
-    )
-
-    object_names_map = get_objects_map([object_names_def_path])
-    game_strings_map = get_strings_map(game_string_def_paths)
-    assert object_names_map
-    assert game_strings_map
-
-    # Update English base files and collect changed pointers
-    changed_ptrs: dict[Path, list[str]] = {}
-    for path in PROJECT_PATHS[game].shipped_data_dir.rglob("*strings*.json*"):
-        # Skip translation files; handle only base (English) files here
-        if is_translation_file(path):
-            continue
-        old_content = path.read_text()
-        new_content = postprocess_game_strings(
-            old_content, object_names_map, game_strings_map
-        )
-        if new_content != old_content:
-            # Write updated English file
-            path.write_text(new_content)
-            # Determine which entries changed for later marking in translations
-            old_data = load_json5_from_string(old_content)
-            new_data = load_json5_from_string(new_content)
-            old_ptr = JSONPointers(old_data)
-            new_ptr = JSONPointers(new_data)
-            changed_ptrs[path] = [
-                ptr
-                for ptr in new_ptr
-                if new_ptr.get(ptr) != old_ptr.get(ptr)
-                and not should_skip_object_name(ptr, new_data)
-            ]
-
-    # Mark changed entries in translation files with REVIEW_MARKER
-    for eng_path, ptrs in changed_ptrs.items():
-        for trans_path in eng_path.parent.glob(f"{eng_path.stem}-*.json5"):
-            if not is_translation_file(trans_path):
-                continue
-            trans_data = load_json5(trans_path)
-            trans_ptr = JSONPointers(trans_data)
-            for ptr in ptrs:
-                trans_ptr[ptr] = REVIEW_MARKER
-            write_json(trans_path, trans_data)
-
-
-def main() -> None:
-    process(game=1)
-    process(game=2)
-
-
-if __name__ == "__main__":
-    main()
